Update parameter default type and value for image shapes
The default image shape datatype for rec_image_shape and cls_image_shape is a list and the default value is also provided as a list argument. However, in the updated version, this is no longer the case.
Now the default type is str, and the default value is also provided by a string. 
Hence, updating these two parameter default datatype and value is required.