Fix "--enable_mkldnn=false" not effective
When using the --enable_mkldnn command line option, MKLDNN should be enabled when set to true and disabled when set to false. But, using the PaddlePaddle v3.0.0-beta1 inference library, the --enable_mkldnn=false option did not work as expected, causing MKLDNN to remain enabled regardless of the setting.
﻿
This commit adds a call to `config.DisableMKLDNN()` in the else branch to explicitly disable MKLDNN when `--enable_mkldnn=false` is set.